Move helper methods out of test files.
