+++ src/testing/test-utils.ts
@@ -1,11 +1,13 @@
 /*
- * Use when working with a jest.fn() mocked method to get better type safety and IDE hinting on
+ * Use when working with a vi.fn() mocked method to get better type safety and IDE hinting on
  * the function signature of what's being mocked.
  *
  * Type "any" is used here to allow for desired type flow during usage where T "looks like a function".
  */
-export const asMockedFn = <T extends (...args: any[]) => any>(fn: T): jest.MockedFunction<T> => {
-  return fn as jest.MockedFunction<T>;
+import { MockedFunction } from 'vitest';
+
+export const asMockedFn = <T extends (...args: any[]) => any>(fn: T): MockedFunction<T> => {
+  return fn as MockedFunction<T>;
 };
 
 export type PromiseController<T> = {
